Update patch set 1
Patch Set 1:

Consider that we use binaries and libraries from a custom root $ROOT. In regular ARMv8 builds, both dalvikvm and dalvikvm64 exist, but only dex2oat (32) exists. To run a dex file, we would use:
$ LD_LIBRARY_PATH=$ROOT/lib64:$ROOT/lib dalvikvm64 ...

Without the patch, this fails, complaining that $ROOT/lib64/lib....so is not 32b. Indeed, dalvikvm64 executes the 32b dex2oat, which thus uses the same LD_LIBRARY_PATH. Since the list starts with a 64b path that contains the required libraries, linker tries to load the first one, and fails since it's not 32b.

With the patch, linker will still fail loading the 64b library, but when this happens it will look at the next path in LD_LIBRARY_PATH, which happens to contain 32b libraries.

Patch-set: 1
Label: Verified=0