Update patch set 1
Patch Set 1: Code-Review+2

> Mostly I'm not a fan of the including of the source from another
 > file.

true, but that's arguably easier to deal with than makefile cleverness. (and we already do it for some of the other tests.)

 > Also, the compiler might choose not to use the fortify
 > version and you don't have that much control.

that's true, but the big hammer suffers from that too. if we fuck up the makefile, we've no way of knowing that we're now uselessly testing the fortified stuff twice. other than getting daily coverage numbers (which is obviously something we want when we have infrastructure), i don't think there's a better alternative?

 > I felt like this was a good compromise to get something done
 > quickly and we can think about a better solution later.

lgtm.

 > I forgot to mention that I tested this on a fugu, put back the bad
 > strrchr, and verified that the regular test passes, but the
 > no_fortify version fails.

ship it!

Patch-set: 1
Label: Code-Review=+2