Update patch set 1
Patch Set 1:

the only reason i remember for this behavior was because glibc's errors (in general, but especially in this situation) really suck. as long as we give good clear errors in the "i'm 32-bit but only have a 64-bit library available" situation (and vice versa), i don't think allowing this is a problem.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0