Update patch set 1
Patch Set 1:

> Does this cause any noticeable slowdown? I don't think we have that
 > much output so it's probably fine, but I could imagine this might
 > run really slowly on something like a watch.
 > 
 > I don't know if there is a faster way to post-process, but if it's
 > really slow maybe there is a way to pre-process.

Going to be nothing measurable in our tests. As you say, there isn't much output, and this also only runs in the event of a failure. Hopefully that's not the normal case :)

Can't pre-process this because the output will always be printed to the console, but might *also* be printed to XML.

Patch-set: 1