Update patch set 2
Patch Set 2:

It does sit as a cts gtest though, that is why this was written, happy to add it, was not sure where to call its home since there are no kernel gTests except as a side-effect in bionic. Decided to move it to system/extras/tests/time/, there will be some cleanup as the existing test may need to turn into gTests as well.

Was not viewed as something we should enforce partners to comply with, functional rtc is not a requirement (!) (my WAG, probably because a vast majority of the Billion devices out there get their time and zone automatically from cell or ntp sources).

The logger attempts some self-correction of (when in realtime mode) timestamps in the transition from rtc to automatic time. Ugly, its not as if the rtc is doing us much of a favour in the automatic world, we would actually be better off without it in those situations because it offers a clearer jump.

Patch-set: 2