Update patch set 2
Patch Set 2:

Sorry, I didn't see but we do have some tests, but they are pretty sparse. Unlike some of the more exhaustive tests that we have for many of the other math functions.

It also appears that one of the math tests for x86 fails (math.lrint), which makes this a risky change to take.

Patch-set: 2
Label: Code-Review=0