Add check for pthread_self() when looking up a thread
Check if thread_id is in fact pthread_self before
locking on g_thread_list_lock in __pthread_internal_find.

This also lets linker use raise() which was not working
because ptrehad_kill() couldn't find pthread_self() thread
because the global thread list is initialized in libc.so and
linkers version of this list is empty.

Bug: http://b/25867917
Change-Id: I18fe620e8cd465b30f0e1ff45fff32958f3c5c00