Update patch set 1
Patch Set 1:

this doesn't change on normal Android systems. there is breakage with Brillo, but there are no systems to maintain compatibility with. I wouldn't mind changing all ports to not have this field and get python to accept it, but we'd rather have all Brillo targets be the same than be similar to ndk warts.

Patch-set: 1
Label: Verified=0