Update patch set 2
Patch Set 2:

> Elliott - We will put together a document with all the Valgrind
 > issues for ARM and ARM64 (once that is ready I will share it with
 > you).
 > 
 > Hopefully this way we can easily track the progress on all the
 > outstanding issues.

thanks, that sounds great. we do have a few people ask now and then about valgrind, but we've largely been pointing them to asan (or x86) lately. would be good to have valgrind as an option again.

(btw, you'll be amused to see that i had to go back to memset because we're still building the pthread code with GCC for mips, and GCC doesn't support = {} :-( )

Patch-set: 2