Update patch set 2
Patch Set 2:

ps2 just adds the bug. which was closed a while back, but i'd kept in a tab because it had annoyed me that we'd just left all this crap lying around.

i do wonder, with bzero out of the way, whether we want _another_ pass over all the __memset_chk implementations so that it falls through to memset (with the branch being for the failure case)...

Patch-set: 2