Update patch set 1
Patch Set 1:

> So as previously discussed with enh, this seems to fail because of unexpected zero value of the canary...

yeah, sorry, i should have follow that suggestion on friday. despite the fact that the other unit test -- the one that checks that if we create a bunch of threads, they all get the same non-zero stack canary value -- is passing, you're right. in the death test, the stack canary TLS value is 0. the global stack guard is still good, though (which explains why the other architectures pass both tests).

the *really* weird part is that the TLS stack guard is already 0 in the death test parent...

Patch-set: 1
Label: Verified=0