Update patch set 10
Patch Set 10:

> > Uploaded patch set 10.
 > 
 > If deprecate int80 completely,  __sysinfo need to be initialized
 > before any system call occur in linker, dynamic libc and static
 > libc.
 > For static case, the initial of __sysinfo should be placed before
 > __libc_inti_main_thread, because __set_tid_address need it.
 > Maybe provide a separated function from __libc_init_globals to do
 > the sysinfo initial is better and place it at the begin of
 > linker_init, __libc_preinit and __libc_init.
 > 
 > bionic_tls.h:
 > #if defined(__i386__)
 > extern __LIBC_HIDDEN__ void* __libc_sysinfo;
 > #endif
 > extern __LIBC_HIDDEN__ void __libc_init_sysinfo(KernelArgumentBlock&);
 > 
 > libc_init_common.cpp:
 > #if defined(__i386__)
 > __LIBC_HIDDEN__ void* __libc_sysinfo = nullptr;
 > #endif
 > 
 > void __libc_init_sysinfo(KernelArgumentBlock& args) {
 > #if defined(__i386__)
 > __libc_auxv = args.auxv;
 > __libc_sysinfo = reinterpret_cast<void*>(getauxval(AT_SYSINFO));
 > #endif
 > }

yeah, sorry, i uploaded this before running for my bus and forgot to comment that it's not finished :-)

you can see the fixes for the dynamic linker are in this change, but -- although i haven't tested yet -- i assume static binaries have the same early RELRO mprotect problem, and need the change like the one you describe here.

the other thing i haven't looked at yet is replacing the function call with the global access.

i also noticed that there's a possibility to fix the current double-initialization of the TLS. if we move the RELRO code out of link_image we can call it after __libc_init_main_thread for the linker, removing the need for the linker to call __set_tls itself. (i might leave that cleanup for dimitry though, since it's not a regression or made worse by this patch.)

Patch-set: 10
Label: Verified=0