Update patch set 11
Patch Set 11:

> More seriously, the ability to boot with vdso32=0 is occasionally
 > useful for testing.

you say that because you're the guy who owns that code on the kernel side :-)

 > And isn't supporting this about one line of
 > code plus an asm helper that is three bytes long?

it's not that i'm worried about; it's about balancing "helpful to the guy who owns the x86 vdso, who's more than smart enough to add those three lines himself if he ever needs to" with "harmful to OEMs who don't really know what they're doing, misconfigure their kernel, and accidentally damage performance because we hide their mistake for them".

my assumption is that the latter scenario is *far* more likely :-(

but if _regular_ folks often need to test without the vdso, that would be different... is there some use case i'm missing?

Patch-set: 11