Add <sys/quota.h>.
It turns out that at least the Nexus 9 kernel is built without CONFIG_QUOTA.
If we decide we're going to mandate quota functionality, I'm happy for us to
be a part of CTS that ensures that happens, but I don't want to be first. For
now, then, I'm happy with ENOSYS. And because I don't have a kernel with
CONFIG_QUOTA, I couldn't go much further even if I wanted to. But the strace
output looks right.

Bug: http://b/27948821
Bug: http://b/27952303
Change-Id: If667195eee849ed17c8fa9110f6b02907fc8fc04