Update patch set 2
Patch Set 2:

> We have no dependency on any specific mapping (we don't care what
 > the final UIDs are), but we do have a dependency on being able to
 > access a range of 1000 UIDs using string names like "oem_XXX". We
 > could convert our stuff to use "oem{1,2}_XXX", so if you want to
 > change oem_XX to be the first range, and oem2_XXX to be the second
 > range, now's a good time to do it. We currently have no hard-coded
 > "oem_XXX" users, but will soon-ish.

given that we have a split range, i'm uncomfortable with having the supposedly user-friendly names (a) backwards and (b) not actually of the form base_offset. (if we're not going to do that, we may as well just call them all "oem_<id>" where id is the raw value.)

oem_xxx and oem2_xxx (instead of the oem1_xxx and oem2_xxx i suggested) is fine with me too, and matches the constants in the code better (unless we want to change them too).

anyway, any of these seem saner to me than trying to awkwardly retrofit into an existing scheme that isn't yet in use :-)

Patch-set: 2
Label: Verified=0