Update patch set 2
Patch Set 2:

> Shouldn't this be fixed on openmp side?

(didn't see this comment in all the treehugger spam.)

that's the question with almost everything we add at this point. in the specific case of OpenMP, yeah, i think it's stupid that they rely on building with this. they should be able to build without it because it's useless bloat. (no end user needs to see OpenMP errors localized, and localization makes it harder for the developer to debug failures from users who speak another language.)

but we have the more general question of "can we fix all the code that's out there, and would it be worth our time even if we could?". 

if you take the extreme version of the former position you wouldn't add anything more to libc and you'd say they should either "fix" the code they're trying to build or statically link in their own implementation. but then that leads to everyone doing our job for us over and over, or us missing out on the things that don't get built/ported.

the extreme opposite position is that we should add everything, no matter how broken or obsolete. even a64l(3). this (over the long term) means that hacks and cruft disappear (or never appear) in third-party code and the only things you can't run are things that are expressly prohibited by SELinux or our kernel configurations.

i agree that <nl_types.h> is a tricky case because it's (a) basically useless and (b) being added only in no-op form. (which is why i resisted for so long.) but it's POSIX, something we can do in half an hour, and sufficient to build OpenMP out of the box. the SysV signal stuff was, i think, a clearer case because it's (a) somewhat useful even if deprecated and (b) used in at least two test suites we want to be able to run. but i don't think we want to get into the project-counting business because we have no idea of what folks are trying to build (sadly), and even if we did we probably don't want to be in the business of deciding whether or not they should be building that, and we definitely don't want to be in the business of saying "you need to rewrite upstream" --- none of us believe that the bionic we inherited was "good enough".

i think the other argument for leaning towards "include" is that we're probably going to need all these things in the NDK compatibility library anyway. (i think we all agree we need the NDK compatibility library, and that its real purpose is "so you can build anything you want and run it on older releases", not "so you can be at the same position as release <X> of Android, because we think that the libc there is good enough for everyone".) and i think we all agree we'd rather have one codebase for that and libc (if only so we have the same bugs everywhere), and even though i think it's going to be hugely useful for many years, long term i'd like to see it vanish. (though there's an alternative position there too --- that the platform is better off the more code that developers statically link because then we can fix bugs at will and not have to worry about existing code that might be relying on bugs. time will tell.)

Patch-set: 2