Update patch set 3
Patch Set 3:

> We will not be moving the Android defaults into /system/etc/passwd
 > anytime soon, and the _full_ list is in-effect, enormous if we
 > include the implied sets (oem_5000 - oem_5999, u0_*). We view these
 > as immutable. When first instantiated, the _only_ ids that will be
 > permitted to land in /system/etc/passwd will be the extension to
 > the oem set (1099 ids in total if all overridden); all others will

I think you should reconsider this, having half-and-half makes the logc more complicated. The libc code needs to check if it is within the oem range check passwd file, else do it differently. Instead of just, check passwd file.

 > be ignored. If we put the defaults in, they will be constructed
 > based on the visible set in the header file, not an _invisible_ set
 > in grp_pwd.cpp ...

That's the point of the library routines getpwduid() and friends. The AID_* are still visible. One can always get the nice name from the uid and vice-versa.

 > 
 > android_ids[], although not part of the official ABI, has a life of
 > its own on partner devices. Until we are sure we can 

Exactly, its private, if you use it, expect it to break with little to no warning. That's the guidance given through the NDK and SDK docs, as well as using private stuff internally in-tree.

move/remove
 > them, let it be, let it be in _one_ place in the include file so we
 > do not maintain several copies.

Now there are just several copies in binaries across the system, and it's still in a header file thus more in tree users will grab it. Perhaps shoving a comment indicating its nearing the end of life and trying to encourage the use of pwd.h/grp.h.

The faster you get android_ids out of the header file, the better IMO.

Patch-set: 3