Update patch set 3
Patch Set 3:

> We will not be moving the Android defaults into /system/etc/passwd
 > anytime soon, and the _full_ list is in-effect, enormous if we
 > include the implied sets (oem_5000 - oem_5999, u0_*). We view these
 > as immutable. When first instantiated, the _only_ ids that will be
 > permitted to land in /system/etc/passwd will be the extension to
 > the oem set (1099 ids in total if all overridden); all others will
 > be ignored. If we put the defaults in, they will be constructed
 > based on the visible set in the header file, not an _invisible_ set
 > in grp_pwd.cpp ...
 > 
 > android_ids[], although not part of the official ABI, has a life of
 > its own on partner devices. Until we are sure we can move/remove
 > them, let it be, let it be in _one_ place in the include file so we
 > do not maintain several copies.

these are two unrelated issues. regardless of whether or not we see the builtins moving into files, i think we should move the array out of system/core. right now there are only four users in vendor/ (all of which are copy & paste) plus one in device/. the sooner you kill those and make it impossible to add more, the better.

all of the existing users can be trivially replaced after the recent bionic changes.

Patch-set: 3