Update patch set 3
Patch Set 3:

partner init files are picked up from /system/etc/init/*.rc; thus /system/etc/passwd is accessible to 'enhance' the oem ids for direct named use in their associated *.rc files for service start. This does not cover early_init stuff of course :-(.

The goals are /system files for the most part are all the same; some files belong to partner like the fs_config files currently generated, and the proposed /system/etc/passwd & groups files. The timing will be a thing of the past once we all move to A/B (there is only a ramdisk for recovery).

Patch-set: 3