Update patch set 5
Patch Set 5: Code-Review+2

(3 comments)

maybe call the other file README_api.md? "libc" seems less obvious.

(also, to be clear: i think we should check in docs that match the behavior now, and if we want to change the behavior in light of this, we can change both together later.)

Patch-set: 5
Label: Code-Review=+2