Update patch set 5
Patch Set 5:

(10 comments)

will need dimitry to confirm that that's the right thing to do with the map. it's a bit of a tricky case where we've had a no-op function for a long time. i suspect the answer is to leave the LP32 stuff as-is, and add to LIBC_O for LP64.

Patch-set: 5