Update patch set 1
Patch Set 1:

It is, but we can not use libc++ (or libbase) and although some std:: templates can be used, one can never guarantee they do not pull in libc++ pieces by accident.

Probably the only C++ism I have here is the pass by reference, C++ comments, and define-where-used for the variables.

ObligatoryDisclaimer: This code may have been 'slapped' together to allow this project to move along, critical and additional code style issues will likely be dealt with on respin. I do not 'like' the general inefficiencies in getpwent/getgrent (reopening passwd/group file _every_ call in the oem aid range). The code is more efficient with a missing file than it is with a content-less file. The group and passwd files are ro and static, so we have a cache opportunity here that was not taken (but we can not use std::unordered_map ...).

Patch-set: 1