Update patch set 2
Patch Set 2:

i think this is readable enough for the NDK headers.

how hard would it be to make the tool remember what the last #if was and merge ones with the same guard? a lot of these would be a lot nicer if we had that, and for some of the others we could just manually group them together. if it's easy, that might be worth doing.

but this is already no worse than most other C libraries!

Patch-set: 2