Enable debug malloc using environment variables
Previously debug malloc can be enabled only using global settings
accessible to the root user only. This CL adds a new option to enable
it using environment variables making it possible to use it with pure
native (shell) applications on production builds (from shell user) and
prepares it for using it from logwrapper on production devices.

Test: manual - enable debug malloc using environment variable
Change-Id: Iecb75a3471552f619f196ad550c5f41fcd9ce8e5