Update patch set 1
Patch Set 1:

> > Thanks to Stephen and Pirama.
 > > The bug fix for this already has been applied to the truck of
 > llvm
 > > (r277501), hence the bug report (https://llvm.org/bugs/show_bug.cgi?id=30308)
 > > is now marked as solved. This commit can be reverted or stayed at
 > > master as it is since NFC.
 > 
 > I think the label attached to the proper code line (instead of the
 > comment) is far easier to read, so I would say not to revert this
 > CL. I do think we should be careful in the future about altering
 > code to work with Clang snapshots. If we can fix the bug in Clang,
 > it really isn't worth checking in most WARs. That said, I view this
 > workaround as a significant readability improvement, so thanks!

the majority of bionic's local labels (whether written by us or outsiders) are at a blank line or one-line comment -- probably because folks use the comment as a replacement for a name -- but this was the only case of a label at a multi-line comment.

it looks like i already rewrote all the other `__bionic_clone`s to use named labels for child and parent, so i'll do that here too.

Patch-set: 1