Remove GCC fortify tests.
The version of GCC we use has internal compiler errors with the
libc++ update. These two tests are the only part of the tree that
trip this issue. The risk here is that we lose test coverage for
fortify with GCC in the NDK. This is pretty low risk since fortify
support doesn't exist at all until the minimum target is KitKat
(Jelly Bean?), and is minimal until Lollipop.

Test: make checkbuild tests
Bug: None
Change-Id: Ifeff414de1d8dfd5d06da0151308e3b5a9daefab