Update patch set 1
Patch Set 1:

The problem happens when following conditions are met.

1) No dedicated vendor partition exists. /vendor/lib/ is symlinked to /system/vendor/lib/.

2) A shared library in the grey list (e.g., libstagefright.so) directly or indirectly requires a proprietary shared library located in /vendor/lib/.

3) That proprietary shared library requires a non-NDK shared library (e.g., libnetutils.so) in /system/lib/.

Currently, is_system_library() returns false for the proprietary shared library because its real path is not /vendor/lib, but /system/vendor/lib. As a result, the non-NDK shared library is wrongly considered to be loaded from a non-system library and thus access is denied.

Patch-set: 1