Update patch set 3
Patch Set 3:

(5 comments)

I could not find a way to set the tid in clone.cpp in a way that worked for all variants of clone (maybe public libc clone should be distinct from an internal routine for fork and pthread_create?). In this patch I go back to invalidating the tid across clone and do a check in gettid().

Patch-set: 3