SafeStack implementation.
This a nearly complete implementation of SafeStack in bionic.
SafeStack allocates a second, "unsafe" stack for every thread and
provides a pointer to that stack through a fixed TLS slot, and also a
simple accessor function (for testing and platforms where the
compiler can not emit direct TLS access yet).

Unsafe stack is mapped with a random sized gap in an attempt to make
offset to the system stack less predictable.

Due to VM pressure SafeStack is disabled on 32-bit targets.

This functionality is controlled by the following build flags:
USE_SAFESTACK=true enables runtime support in bionic,
SANITIZE_TARGET=safe-stack enables compiler support (code
  instrumentation).

Libc and even the linker itself are compiled with SafeStack
instrumentation (depending on the SANITIZE_TARGET setting), with the
exception of two tiny functions (one in linker, one in
libc_init_static) that are used to bootstrap the system.

Test: no functional change by default (unless USE_SAFESTACK is set)

Change-Id: Id82fc4571134c25710f7d228f1b11b15870312d7