Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> why isn't this a problem for all accessors of `tid`? https://cs.corp.google.com/search/?q=%5C-%3E%5Cs*%5Cbtid%5Cb+file:bionic/&m=100&sort=1&sq=package:%5Eandroid$&type=cs

It could be for any single-threaded program that calls clone() directly (as opposed of through fork()) and then does not exec(). (I'm treating multi-threaded clone like a multi-threaded fork in terms of async-signal-safety requirements).

Some of the other pthread calls I could switch to also using gettid(), but things that use __pthread_internal_find() could still operate on a pthread with an invalid tid.

The other option would be to fix-up the tid in __start_thread. I've posted patch set 2 which does that for discussion, as an alternative to updating all the other pthread_internal_t::tid accesses.

Patch-set: 1
Label: Verified=0