Sigaltstack for SafeStack.
Linux kernel does not implement sigaltstack switching for the unsafe
stack, so we have to do it in userspace. This change wraps on ONSTACK
signal handlers in a function that swaps the unsafe stack. The
altstack is always allocated by libc and has the same size as the
user-provided sigaltstack.

Test: bionic device tests

Bug: 32208273
Change-Id: I24e71761c6840f63d9fb6c763b1bc35af2f34546