Bionic linker is no longer hijacking libdl.so
Do not hijack libdl.so methods but make libdl proxy calls to
linker instead. This will be replaces by calls to libc.so
onces linker functionality is migrated.

As a side effect this commit makes it possible to intercept
calls to dlopen/dlsym and co.

Also add a lock on dl_unwind_find_exidx function call.

Test: bionic-unit-tests --gtest_filter=dl*:Dl*
Change-Id: I9e666e771e4bbca52151cfa7fc4c8677e1480818