Update patch set 2
Patch Set 2: Verified+1

actually, when i went to look at what needed to be cleaned up, i realized that the problem was actually that our <limits.h> always #included <sys/limits.h> anyway. so the net effect of this change is to make <sys/limits.h> no longer a *subset* of the POSIX <limits.h> but exactly the same (and to remove the questionable copy & paste of <float.h> that was in <sys/limits.h>).

so not the outcome i was hoping for, but an improvement anyway .

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1