Update patch set 9
Patch Set 9:

I've not rebased it in a couple of months, but it looks good.
It is more or less complete now (diagnostics support is in).
Build system changes are already in.
I think current clang toolchain misses one important change, but that should not block this review.

Patch-set: 9
Label: Verified=0