Add an additional offset to zoneinfo
Add an additional offset to the zoneinfo header to allow for
future evolution of the format and allow detection
of file truncation.

Prior releases of Android will ignore the new offset
and continue to assume the zoneTab section extends
until the EOF. This will allow Android to use the same
zoneinfo files across versions until more sections
(which is not currently planned).

The new offset is completely ignored by bionic because it
already ignored the zoneTab section.

The tzdata file has been updated to incorporate the new
header field.

Bug: 31008728
Test: Manual testing / CtsBionicTestCases
Change-Id: I5401ecb7866f3827d072a124cfb9689edc23a08d