Update patch set 6
Patch Set 6: Code-Review+2

There is a small regression for memset on the big core on a nexus 5x for sizes of about 144 to 256. However, this is relatively tiny and the boost for the small core is big enough that I don't think this matters.

Here are the configurations I benchmarked:

Pixel
Nexus 5x (small core)
Nexus 5x (big core)

I also ran the bionic unit tests in all of these configurations:

Pixel
Nexus 5x
Nexus 6p

So it all looks good to me.

Patch-set: 6
Label: Code-Review=+2