WIP: Relax name/value length limits on system properties
This change allows name and value lengths of a system property to be
as long as 2^9 characters (including nul-terminator).

Following new APIs are added:
__system_property_get_string
__system_property_read_name
__system_property_read_value

These APIs are guaranteed to return name/value without any truncation.

Existing following APIs give truncated result when the length exceeds
old limit (32 for name, 92 for value):

__system_property_get
__system_property_read

Internally, prop_info structure is modified so that values are stored in
a dynamically allocated array. When a sysprop is created or updated, an
array of appropriate size is created and the value string is stored
there.

Support for the old prop_info structure is done as a compatibility mode.
This mode is required when static binary linking this new bionic runs on
devices with old bionic (during OTA).

Finally, code for the old compatibility mode has been removed because it
is for OTA from KitKat devices, which we don't seem to support anymore.

Bug: 33926793
Test: should be able to set/get sysprops within the new limit 2^9

Change-Id: I77be98d596720034fabf21528036a44a8fb212c9